requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
watchdog==4.0.0
anthropic==0.18.1
python-dotenv==1.0.1
